acpi: Remove netlink based propagation of power button events
It is wrong to unconditionally propagate ACPI power button events in the
host to the guest, since some guests (e.g. ARCVM) don't expect them and
misbehave when such an event is received. If at all, this propagation
should be enabled only in crosvm builds with "direct" feature enabled,
i.e. crosvm-direct builds for ManaTEE.

Moreover, even for ManaTEE, the correct approach is to propagate them
via direct forwarding of physical ACPI events, rather than via listening
on ACPI netlink events in the host (in particular because we want them
to be handled in the guest only, not in the host).

Since CL:3645408 and CL:3599200 we are using direct forwarding of ACPI
power button events, so the netlink based forwarding in the ACPI event
listener is no more used by anyone and can be removed.

BUG=b:241191471, b:230592368
TEST=ACPI power button events are not propagated to ARCVM guest.

Change-Id: I2541bf900d51507dc246b46fb6691dfa0b1c2956
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3859626
Tested-by: Dmytro Maluka <dmaluka@google.com>
Reviewed-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Dmytro Maluka <dmaluka@google.com>