acpi: Remove netlink based propagation of GPE events
Currently both direct and non-direct GPEs are unconditionally forwarded
to the guest, just in different ways. This is wrong not only because we
don't necessarily want to propagate the given non-direct GPE to the
guest, but also because non-direct GPEs are forwarded without preventing
handling them in the host, so that a GPE handler is run twice: in the
host and in the guest. This behavior doesn't seem to make sense and is
generally buggy.

The idea of this patch is that acpi device code should forward direct
GPEs only, and for non-direct GPEs it should just notify the registered
gpe_notify listeners, if any. Those listeners may then inject the given
GPE to the guest or do something else, depending on the purpose of the
given GPE.

This patch doesn't change any behavior for direct GPEs, so currently it
doesn't affect ManaTEE, since currently all 4 GPEs on Brya/Volteer are
used as direct GPEs. However in the future we may use non-direct GPEs
e.g. for a cleaner solution for handling Thunderbolt hotplug GPE.

BUG=b:230592368
TEST=Inspect /sys/firmware/acpi/interrupts/ to see if GPE events (EC,
Thunderbolt etc) work correctly in ManaTEE. No specific test for
non-direct GPEs, since we don't use them.

Change-Id: I4c9b59c217ce347cfe06decf25694cbf6c4170f4
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3861973
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Dmytro Maluka <dmaluka@google.com>
Tested-by: Dmytro Maluka <dmaluka@google.com>