devices: virtio: block: remove Default for DiskOption
The Default impl for DiskOption was only used in one place (vhost-user
block device); that use can be replaced by filling in the two missing
fields.

The values used in the Default impl did not match the serde deserializer
defaults, which are used when parsing the --disk and related options
from the command line. Removing this unnecessary Default clarifies which
values are actually used.

BUG=b:237829580
TEST=tools/presubmit --all

Change-Id: I5ffaa4facab31402aeb84552cadce750d132c307
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3877202
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>