video: decoder: vaapi: do not flush if the submit queue is not empty
Do not flush the codec if the submit queue is not empty. Flushing the
codec backend will release the reference frames in most (all) codecs.

This cannot happen if we still have pending frames in the submit queue
as these frames would not find their references, breaking the decoding
process.

Not only that, but as the length of the submit/ready queues is
runtime-dependent, we would end up with a random number of corrupted
frames during flush.

Instead, signal that we want to flush and try making progress
accordingly, but only flush the backend once it has processed all the
pending frames in the submit queue.

BUG=b:214478588
TEST=test-25fps.h264 decodes without corruption in the last (flushed)
frames

Change-Id: I4200102344bab86c9938ae898383bb82e6542594
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3907530
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>