serde_keyvalue: make unquoted strings stop at the next bracket
In order to support tuples and enum variants, we need to add an extra
separator to specify the bounds of the tuple or enum. `[` and `]` are
good candidates since they are not often met in strings and are not
reserved by the shell, contrary to `(` and `)`.

BUG=b:248993755
TEST=cargo test -p serde_keyvalue
TEST=cargo test

Change-Id: I6743461f0cd3af0bce11e4978fa14776048170a5
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3915038
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>