media: ffmpeg: Add helper to create AvPixelFormat from raw enum values.
Since re-enumerating all the formats supported by FFmpeg in our bindings
would be a rather large effort, it looks like a good compromise here is
to allow the user to create an AvPixelFormat from the raw enum values
exposed by bindgen.

Add a helper that checks that the value is in range of valid formats and
convert it into our enum wrapper.

BUG=b:239897269
TEST=cargo test --features ffmpeg -p ffmpeg

Change-Id: I3c1433e9d76d645b7615bde81433e1df2cfa3910
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3911108
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>