media: cros-codecs: decoders: allow for polling the decoders
Allow for polling the decoders. This will return any pending decode
requests that happen to be ready. If block is set to true, it will
block until they are all ready.

This is different from flush, in that flush will empty the decoder's
internal state, altering the DPB, whereas poll will not.

BUG=b:214478588
TEST=cd media/cros-codecs && cargo test --features vaapi

Change-Id: I56b1b9d7b704c17d75bbe93d4fd4eb0a9a468af7
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3918838
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>